+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Wed, 09 Feb 2022 20:14:50 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Feb 2022 22:48:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.27.30
+
+Tue, 15 Feb 2022 22:48:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.22.0
+- Bump @fluentui-react-native/menu-button to v0.7.30
+
+## 0.27.29
+
+Mon, 14 Feb 2022 23:29:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.16.2
+- Bump @fluentui-react-native/menu-button to v0.7.29
+
+## 0.27.28
+
+Thu, 10 Feb 2022 17:13:16 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.24
+- Bump @fluentui-react-native/contextual-menu to v0.16.1
+- Bump @fluentui-react-native/menu-button to v0.7.28
 
 ## 0.27.27